gpu-requirements.txt
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==4.4.2
easydict==1.13
eval_type_backport==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
insightface==0.7.3
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime-gpu==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
prettytable==3.11.0
proglog==0.1.10
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==2.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
